Add operations for new stocktake API with storeID as a parameter and newly named fields
