+++ packages/common/src/operations.graphql
@@ -1,121 +1,3 @@
-query stocktake($stocktakeId: String!, $storeId: String!) {
-  stocktake(id: $stocktakeId, storeId: $storeId) {
-    __typename
-    ... on StocktakeNode {
-      __typename
-      id
-      stocktakeNumber
-      comment
-      createdDatetime
-      # stocktakeDatetime
-      status
-      description
-      # entryDatetime
-      # enteredByName
-      # onHold
-      lines {
-        __typename
-
-        ... on StocktakeLineConnector {
-          __typename
-          nodes {
-            __typename
-            batch
-            # itemCode
-            # itemName
-            itemId
-            id
-            expiryDate
-            packSize
-            snapshotNumberOfPacks
-            countedNumberOfPacks
-            sellPricePerPack
-            costPricePerPack
-          }
-          totalCount
-        }
-      }
-    }
-  }
-}
-
-mutation upsertStocktakeLines(
-  $storeId: String!
-  $deleteStocktakeLines: [DeleteStocktakeLineInput!]
-  $updateStocktakeLines: [UpdateStocktakeLineInput!]
-  $insertStocktakeLines: [InsertStocktakeLineInput!]
-) {
-  batchStocktake(
-    storeId: $storeId
-    input: {
-      deleteStocktakeLines: $deleteStocktakeLines
-      updateStocktakeLines: $updateStocktakeLines
-      insertStocktakeLines: $insertStocktakeLines
-    }
-  ) {
-    __typename
-    ... on BatchStocktakeResponses {
-      __typename
-      deleteStocktakeLines {
-        id
-      }
-      insertStocktakeLines {
-        id
-      }
-      updateStocktakeLines {
-        id
-      }
-    }
-  }
-}
-
-# query stocktakes($params: StocktakeListParameters) {
-#   stocktakes(params: $params) {
-#     __typename
-#     ... on StocktakeConnector {
-#       nodes {
-#         id
-#         comment
-#         description
-#         stocktakeDatetime
-#         stocktakeNumber
-#         status
-#       }
-#       totalCount
-#     }
-#   }
-# }
-
-mutation deleteStocktakes($storeId: String!, $ids: [DeleteStocktakeInput!]) {
-  batchStocktake(storeId: $storeId, input: { deleteStocktakes: $ids }) {
-    __typename
-    ... on BatchStocktakeResponses {
-      deleteStocktakes {
-        __typename
-        id
-      }
-    }
-  }
-}
-
-mutation updateStocktake($input: UpdateStocktakeInput!) {
-  updateStocktake(input: $input) {
-    ... on StocktakeNode {
-      __typename
-      id
-    }
-  }
-}
-
-mutation insertStocktake($input: InsertStocktakeInput!) {
-  insertStocktake(input: $input) {
-    ... on StocktakeNode {
-      __typename
-      id
-    }
-  }
-}
-
 query names(
   $key: NameSortFieldInput!
   $desc: Boolean
